+++ src/requirements.txt
@@ -1,0 +1,8 @@
+python-dotenv
+tdqm
+numpy==1.19.5
+nltk
+flask
+flask_cors
+tensorflow>=2
+sklearn
