+++ src/src/scripts/train_rnn.py
